// condoplanner-app/src/apiClient/models/EStatusReserva.ts
/* tslint:disable */
/* eslint-disable */
/**
 * API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const EStatusReserva = {
    Pendente: 'PENDENTE',
    Aprovado: 'APROVADO',
    Rejeitado: 'REJEITADO',
    Cancelado: 'CANCELADO',
    Concluida: 'CONCLUIDA'
} as const;
export type EStatusReserva = typeof EStatusReserva[keyof typeof EStatusReserva];


export function instanceOfEStatusReserva(value: any): boolean {
    for (const key in EStatusReserva) {
        if (Object.prototype.hasOwnProperty.call(EStatusReserva, key)) {
            if (EStatusReserva[key as keyof typeof EStatusReserva] === value) {
                return true;
            }
        }
    }
    return false;
}

export function EStatusReservaFromJSON(json: any): EStatusReserva {
    return EStatusReservaFromJSONTyped(json, false);
}

export function EStatusReservaFromJSONTyped(json: any, ignoreDiscriminator: boolean): EStatusReserva {
    return json as EStatusReserva;
}

export function EStatusReservaToJSON(value?: EStatusReserva | null): any {
    return value as any;
}

export function EStatusReservaToJSONTyped(value: any, ignoreDiscriminator: boolean): EStatusReserva {
    return value as EStatusReserva;
}